+++ players/view/process.php
@@ -9,6 +9,7 @@ if(!empty($_POST))
   require_once("../../classes/class.Player.inc.php");
 
 	$player = new \WurmUnlimitedAdmin\PLAYER();
+	$_POST["gmUserName"] = $_SESSION["userData"]["username"];
 	switch($_POST["doing"])
 	{
 		case "banFunction":